fix mistakes after mentor review
