+++ src/main/java/mate/academy/bookstore/controller/AuthenticationController.java
@@ -2,8 +2,8 @@ package mate.academy.bookstore.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import mate.academy.bookstore.dto.UserRegistrationRequestDto;
-import mate.academy.bookstore.dto.UserResponseDto;
+import mate.academy.bookstore.dto.user.UserRegistrationRequestDto;
+import mate.academy.bookstore.dto.user.UserResponseDto;
 import mate.academy.bookstore.exception.RegistrationException;
 import mate.academy.bookstore.service.UserService;
 import org.springframework.web.bind.annotation.PostMapping;
