+++ src/main/java/mate/academy/bookstore/service/UserService.java
@@ -1,7 +1,7 @@
 package mate.academy.bookstore.service;
 
-import mate.academy.bookstore.dto.UserRegistrationRequestDto;
-import mate.academy.bookstore.dto.UserResponseDto;
+import mate.academy.bookstore.dto.user.UserRegistrationRequestDto;
+import mate.academy.bookstore.dto.user.UserResponseDto;
 
 public interface UserService {
     UserResponseDto register(UserRegistrationRequestDto requestDto);
